backend/requirements.txt
# Django Core
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & JWT
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Security & Encryption
cryptography==41.0.7
bcrypt==4.1.2
django-ratelimit==4.1.0

# Two Factor Authentication
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0

# Database
psycopg==3.1.13

# Environment Variables
python-decouple==3.8

# Development
django-debug-toolbar==4.2.0
flake8==6.1.0
black==23.11.0

# Documentation
drf-spectacular==0.26.5

# Additional utilities
requests==2.31.0
python-dateutil==2.8.2
